# docs/swagger.yml
openapi: 3.0.3
info:
  title: Financial Assistance Application API
  version: 1.0.0
  description: This API enables the management and processing of financial assistance applications.
    It allows students to apply for grants, and provides system admins and other roles access to manage and track the status of applications.
    The API supports roles such as students, financial assistance coordinators, and various offices responsible for managing the budget, processing payments, and approving grants.
    Key features include application submission, status updates, and administrative control over grant approvals and disbursements.
servers:
  - url: 'https://{environment}/api/v1'
    description: Server url
...
paths:
...
  /students:
    # post:
    #   tags:
    #     - Student
    #   summary: Add student information
    #   requestBody:
    #     required: true
    #     content:
    #       application/json:
    #         schema:
    #           $ref: '#/components/schemas/Student'
    #   responses:
    #     '200':
    #       description: Login successful
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/StudentCreateSuccessResponse'
    #     '400':
    #       description: Invalid request
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/ErrorResponse'
    #     '403':
    #       description: Invalid request
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/ForbiddenErrorResponse'
    #     '500':
    #       description: Invalid request
    #       content:
    #         application/json:
    #           schema:
    #             $ref: '#/components/schemas/InternalErrorResponse'
    get:
      tags:
        - Student
      summary: Get all students
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search term for filtering users
        - in: query
          name: sort
          schema:
            type: string
          description: Field to sort by
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
          description: Number of records to skip
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          description: Maximum number of records to return
      responses:
        '200':
          description: A list of students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllStudentSuccessResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /students/{userId}:
    get:
      tags:
        - Student
      summary: Get a student
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user with student details
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      responses:
        '200':
          description: A list of students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStudentSuccessResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /students/{studentId}:
    put:
      tags:
        - Student
      summary: Edit student information
      description: Updates the details of an existing student.
      parameters:
        - name: studentId
          in: path
          required: true
          description: The ID of the student to be updated
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentUpdateSuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
    delete:
      tags:
        - Student
      summary: Delete a student record
      description: Delete the details of an existing student.
      parameters:
        - name: studentId
          in: path
          required: true
          description: The ID of the student to be deleted
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      responses:
        '200':
          description: delete details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessDeleteResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          example: INVALID-REQUEST
        generatedAt:
          type: integer
          example: 1738335047588
        description:
          type: string
          example: Request is not well-formed, syntactically incorrect, or violates schema.
        errorDetails:
          type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    example: field
                  value:
                    type: string
                    example: qwer2@gmail.com
                  msg:
                    type: string
                    example: E-mail already in use
                  path:
                    type: string
                    example: username
                  location:
                    type: string
                    example: body
...
    InternalErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          example: INTERNAL-SERVER-ERror
        generatedAt:
          type: integer
          example: 1738335047588
        description:
          type: string
          example: Server Errror Occured
        errorDetails:
          type: string
          example: Server Errror Occured
    ForbiddenErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          example: FORBIDDEN
        generatedAt:
          type: integer
          example: 1738335047588
        description:
          type: string
          example: The resource requested is forbidden for your API Key.
        errorDetails:
          type: string
          example: Invalid API Token
    NotFoundErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          example: NOT-FOUND
        generatedAt:
          type: integer
          example: 1738335047588
        description:
          type: string
          example: The resource requested is not found.
        errorDetails:
          type: string
          example: 404 not found
...
    Student:
      type: object
      properties:
        studentId:
          type: string
          example: 7ec6d79b-ad78-4d19-996b-9e24996e13f7
        firstName:
          type: string
          example: John
        middleName:
          type: string
          example: Michael
        lastName:
          type: string
          example: Doe
        extension:
          type: string
          example: Jr
        sex:
          type: string
          example: Male
        placeOfBirth:
          type: string
          example: New York, USA
        birthdate:
          type: string
          format: date
          example: '2000-01-01'
        height:
          type: integer
          example: 170
        weight:
          type: integer
          example: 70
        permanentStreet: 
          type: string
          example: 123 Main Street
        permanentBarangayId: 
          type: integer
          example: 170
        permanentCityMunId: 
          type: integer
          example: 170
        permanentProvinceId:
          type: integer
          example: 170
        permanentCountry:
          type: string
          example: Philippines
        permanentZipCode:
          type: string
          example: 9000
        currentStreet: 
          type: string
          example: 123 Main Street
        currentBarangayId: 
          type: integer
          example: 170
        currentCityId: 
          type: integer
          example: 170
        currentProvinceId:
          type: integer
          example: 170
        currentCountry:
          type: string
          example: Philippines
        currentZipCode:
          type: string
          example: 9000
        email:
          type: string
          example: john.doe123@example.com
        mobileNumber:
          type: string
          example: '+1234567890'
        soloParent:
          type: boolean
          example: false
        childOfSoloParent:
          type: boolean
          example: false
        indigenousMember:
          type: boolean
          example: true
        indigenousGroup:
          type: string
          example: Cherokee
        spEd:
          type: boolean
          example: false
        pwd:
          type: boolean
          example: false
        grade12SchoolId:
          type: string
          example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
        grade12Strand:
          type: string
          example: STEM
        grade12ProgramName:
          type: string
          example: Default
        grade12Award:
          type: string
          example: Valivecturian
        grade12Organization:
          type: string
          example: APS
        grade12YearOfGraduation:
          type: integer
          example: 2018
        collegeProgramName:
          type: string
          example: Bachelor of Science in Computer Science
        collegeYearLevel:
          type: integer
          example: 2
        collegeAwardHonor:
          type: string
          example: Dean's Lister
        collegeOrganization:
          type: string
          example: Computer Society
        collegeSchoolId:
          type: string
          example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
        emergencyFullName:
          type: string
          example: Jane Doe
        emergencyMobileNumber:
          type: string
          example: '+639938172631'
        emergencyRelationship:
          type: string
          example: Sister
        fatherFirstName:
          type: string
          example: Michael
        fatherMiddleName:
          type: string
          example: Alan
        fatherLastName:
          type: string
          example: Doe
        fatherExtension:
          type: string
          example: Sr
        fatherOccupation:
          type: string
          example: Engineer
        fatherIncome:
          type: integer
          example: 80000
        fatherMobileNumber:
          type: string
          example: '+639938172631'
        motherFirstName:
          type: string
          example: Maria
        motherMiddleName:
          type: string
          example: Therese
        motherLastName:
          type: string
          example: Smith
        motherExtension:
          type: string
          example: ''
        motherOccupation:
          type: string
          example: Teacher
        motherIncome:
          type: integer
          example: 60000
        motherMobileNumber:
          type: string
          example: '+639938172631'
        guardianFirstName:
          type: string
          example: Anna
        guardianMiddleName:
          type: string
          example: Grace
        guardianLastName:
          type: string
          example: Johnson
        guardianExtension:
          type: string
          example: ''
        guardianOccupation:
          type: string
          example: Nurse
        guardianIncome:
          type: integer
          example: 50000
        guardianMobileNumber:
          type: string
          example: '+639938172631'
        siblings:
          type: array
          items:
            type: object
            properties:
              siblingName:
                type: string
                example: Emile
              siblingBirthdate:
                type: string
                format: date
                example: '2005-02-15'
              siblingAge:
                type: integer
                example: 19
              siblingStatus:
                type: string
                example: Single
              siblingRemarks:
                type: string
                example: Studying
              livingWithParents:
                type: boolean
                example: true
              ownHouse:
                type: boolean
                example: false
...
    StudentCreateSuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Resource Successfully Updated!
        data:
          $ref: '#/components/schemas/Student'
    StudentUpdateSuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Resource Successfully Updated!
        data:
          $ref: '#/components/schemas/Student'
        generatedAt:
          type: integer
          example: 1738335634108
    GetAllStudentSuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Resource Successfully Retrieved!
        data:
          type: object
          properties:
            count:
              type: number
              example: 1
            students:
              type: array
              items:
                $ref: '#/components/schemas/Student'
        generatedAt:
          type: integer
          example: 1738335634108
    GetStudentSuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Resource Successfully Retrieved!
        data:
          $ref: '#/components/schemas/Student'
        generatedAt:
          type: integer
          example: 1738335634108
...
    SuccessDeleteResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Resource Successfully Deleted!
...